refactor: merge deploy and data replication tests
Consolidate deploy_replication_test.go into data_replication_test.go to
avoid duplicate primary+replica deployments. CNPG config assertions now
run in the BeforeAll of the data replication spec, cutting total test
time by eliminating a redundant ~2 min setup phase.

Move findCNPGCluster() helper to helpers_test.go so it is shared by both
the data replication and failover test files.
